Refactor options UI
