+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2025-07-30
 
+- 1501 Refactored OptionsUI by extracting item preview functionality into a separate module called ItemPreview. The new class encapsulates preview initialization, event handling, animation control, and item rendering. OptionsUI now delegates preview setup to this class, trimming over one hundred lines from the file. This structural change improves readability and makes it easier to maintain or extend each feature independently. Existing behavior for toggling the preview remains consistent, but the codebase is now more modular.
+
 - 1418 Updated amphitheater seating collisions so that each seat row segment is fully collidable and matches the stage platform's solidity. Added radial collision parameters to individual segments to correctly handle walking on curved tiers, preventing the player from falling through or clipping when traversing seats. Seat bases now register as barriers with stair properties so characters can smoothly step onto them while still interacting with prompts. These changes align seating behavior with the stage and provide a more consistent exploration experience throughout the amphitheater.
